It's not an int64_t anymore (it was,  in an older version of this code).

+++ ades2mpc.cpp
@@ -50,7 +50,7 @@ encountering closing tags).  */
 int free_ades2mpc_context( void *context)
 {
    ades2mpc_t *tptr = (ades2mpc_t *)context;
-   const int64_t rval = tptr->depth;
+   const int rval = tptr->depth;
 
    free( context);
    return( rval);
